feat: initialize server and database connection
feat: set up Express app with middleware for security, logging, and error handling

feat: implement CORS configuration for development and production environments

feat: create MongoDB connection utility with error handling and connection state management

feat: add user authentication controller with registration and login functionality

feat: implement lead management controller for creating and retrieving leads

feat: add user management controller for retrieving and updating user information

feat: create middleware for authentication and authorization checks

feat: implement database connection check middleware for API requests

feat: add global error handling middleware for consistent error responses

feat: create validation middleware for user and lead data

feat: define Mongoose models for User and Lead with validation rules

feat: set up API routes for authentication, user management, and lead management

chore: add Vercel configuration for deployment
