+++ package.json
@@ -2,10 +2,10 @@
   "name": "da-backend",
   "version": "1.0.0",
   "description": "Express backend server for DA Admin",
-  "main": "index.js",
+  "main": "server.js",
   "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js",
+    "start": "node server.js",
+    "dev": "nodemon server.js",
     "build": "echo 'No build step required'",
     "vercel-build": "echo 'Building for Vercel'"
   },
