--- routes/api.js
@@ -1,738 +1,0 @@
-const express = require('express');
-const jwt = require('jsonwebtoken');
-const { body, validationResult } = require('express-validator');
-const mongoose = require('mongoose');
-const router = express.Router();
-
-const User = require('../models/User');
-const Lead = require('../models/Lead');
-const { authMiddleware, adminMiddleware } = require('../middleware/auth');
-
-// Helper function to generate JWT token
-const generateToken = (userId) => {
-  if (!process.env.JWT_SECRET) {
-    throw new Error('JWT_SECRET environment variable is not defined');
-  }
-  return jwt.sign({ userId }, process.env.JWT_SECRET, {
-    expiresIn: '7d' // Token expires in 7 days
-  });
-};
-
-// Auth routes
-const authRouter = express.Router();
-
-// Register/Signup endpoint
-authRouter.post('/register', [
-  body('name')
-    .trim()
-    .isLength({ min: 2, max: 50 })
-    .withMessage('Name must be between 2 and 50 characters'),
-  body('email')
-    .isEmail()
-    .normalizeEmail()
-    .withMessage('Please enter a valid email'),
-  body('password')
-    .isLength({ min: 6 })
-    .withMessage('Password must be at least 6 characters long')
-], async (req, res) => {
-  try {
-    // Check if database is connected (allow connecting state)
-    if (mongoose.connection.readyState === 0 || mongoose.connection.readyState === 3) {
-      console.error('🔥 Database connection state:', mongoose.connection.readyState);
-      console.error('🔥 Available states: 0=disconnected, 1=connected, 2=connecting, 3=disconnecting');
-      return res.status(503).json({
-        success: false,
-        message: 'Database connection not available',
-        debug: {
-          connectionState: mongoose.connection.readyState,
-          hasMongoUri: !!(process.env.MONGODB_URI || process.env.DA_DATABASE_URL_MONGODB_URI)
-        }
-      });
-    }
-    
-    // If connecting (state 2), wait a moment for connection to complete
-    if (mongoose.connection.readyState === 2) {
-      console.log('⏳ Database is connecting, waiting...');
-      await new Promise(resolve => setTimeout(resolve, 2000)); // Wait 2 seconds
-      
-      if (mongoose.connection.readyState !== 1) {
-        return res.status(503).json({
-          success: false,
-          message: 'Database connection timeout',
-          debug: {
-            connectionState: mongoose.connection.readyState,
-            hasMongoUri: !!(process.env.MONGODB_URI || process.env.DA_DATABASE_URL_MONGODB_URI)
-          }
-        });
-      }
-    }
-    
-    // Check for validation errors
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        success: false,
-        message: 'Validation failed',
-        errors: errors.array()
-      });
-    }
-
-    const { name, email, password } = req.body;
-
-    // Check if user already exists
-    let existingUser;
-    try {
-      existingUser = await User.findByEmail(email);
-    } catch (findError) {
-      console.error('Error checking existing user:', findError);
-      return res.status(500).json({
-        success: false,
-        message: 'Database error while checking user existence'
-      });
-    }
-    
-    if (existingUser) {
-      return res.status(400).json({
-        success: false,
-        message: 'User with this email already exists'
-      });
-    }
-
-    // Create new user
-    const user = new User({
-      name,
-      email,
-      password
-    });
-
-    // Save user to database
-    const savedUser = await user.save();
-
-    // Generate JWT token
-    const token = generateToken(savedUser._id);
-
-    res.status(201).json({
-      success: true,
-      message: 'User registered successfully',
-      token,
-      user: {
-        id: savedUser._id,
-        name: savedUser.name,
-        email: savedUser.email,
-        role: savedUser.role,
-        createdAt: savedUser.createdAt
-      }
-    });
-
-  } catch (error) {
-    console.error('Registration error:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Server error during registration'
-    });
-  }
-});
-
-// Login endpoint
-authRouter.post('/login', [
-  body('email')
-    .isEmail()
-    .normalizeEmail()
-    .withMessage('Please enter a valid email'),
-  body('password')
-    .notEmpty()
-    .withMessage('Password is required')
-], async (req, res) => {
-  try {
-    // Check for validation errors
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        success: false,
-        message: 'Validation failed',
-        errors: errors.array()
-      });
-    }
-
-    const { email, password } = req.body;
-
-    // Find user by email and include password for comparison
-    const user = await User.findByEmail(email).select('+password');
-    if (!user) {
-      return res.status(401).json({
-        success: false,
-        message: 'Invalid email or password'
-      });
-    }
-
-    // Check if user is active
-    if (!user.isActive) {
-      return res.status(401).json({
-        success: false,
-        message: 'Account is deactivated. Please contact support.'
-      });
-    }
-
-    // Compare password
-    const isPasswordValid = await user.comparePassword(password);
-    if (!isPasswordValid) {
-      return res.status(401).json({
-        success: false,
-        message: 'Invalid email or password'
-      });
-    }
-
-    // Update last login
-    user.lastLogin = new Date();
-    await user.save();
-
-    // Generate JWT token
-    const token = generateToken(user._id);
-
-    res.json({
-      success: true,
-      message: 'Login successful',
-      token,
-      user: {
-        id: user._id,
-        name: user.name,
-        email: user.email,
-        role: user.role,
-        lastLogin: user.lastLogin
-      }
-    });
-
-  } catch (error) {
-    console.error('Login error:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Server error during login'
-    });
-  }
-});
-
-// Get current user profile (protected route)
-authRouter.get('/me', authMiddleware, async (req, res) => {
-  try {
-    res.json({
-      success: true,
-      user: req.user
-    });
-  } catch (error) {
-    console.error('Get profile error:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Server error fetching profile'
-    });
-  }
-});
-
-// Logout endpoint (optional - mainly for client-side token removal)
-authRouter.post('/logout', authMiddleware, (req, res) => {
-  res.json({
-    success: true,
-    message: 'Logged out successfully'
-  });
-});
-
-// Users routes (protected)
-const usersRouter = express.Router();
-
-// Get all users (admin only)
-usersRouter.get('/', authMiddleware, adminMiddleware, async (req, res) => {
-  try {
-    // Check database connection
-    if (mongoose.connection.readyState !== 1) {
-      return res.status(503).json({
-        success: false,
-        message: 'Database connection not available'
-      });
-    }
-
-    const page = parseInt(req.query.page) || 1;
-    const limit = parseInt(req.query.limit) || 10;
-    const search = req.query.search;
-    const role = req.query.role;
-
-    // Build query
-    let query = {};
-    if (role && role !== 'all') {
-      query.role = role;
-    }
-    if (search) {
-      query.$or = [
-        { name: { $regex: search, $options: 'i' } },
-        { email: { $regex: search, $options: 'i' } }
-      ];
-    }
-
-    // Get users with pagination
-    const users = await User.find(query)
-      .select('-password') // Exclude password field
-      .sort({ createdAt: -1 })
-      .skip((page - 1) * limit)
-      .limit(limit);
-
-    // Get total count for pagination
-    const total = await User.countDocuments(query);
-
-    // Get role counts for dashboard
-    const roleCounts = await User.aggregate([
-      { $group: { _id: '$role', count: { $sum: 1 } } }
-    ]);
-
-    const roleStats = roleCounts.reduce((acc, item) => {
-      acc[item._id] = item.count;
-      return acc;
-    }, {});
-
-    res.json({
-      success: true,
-      data: {
-        users,
-        pagination: {
-          page,
-          limit,
-          total,
-          pages: Math.ceil(total / limit)
-        },
-        stats: {
-          total,
-          ...roleStats
-        }
-      }
-    });
-
-  } catch (error) {
-    console.error('❌ Error fetching users:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Error fetching users',
-      error: process.env.NODE_ENV === 'development' ? error.message : 'Internal server error'
-    });
-  }
-});
-
-// Get user by ID (admin only)
-usersRouter.get('/:id', authMiddleware, adminMiddleware, async (req, res) => {
-  try {
-    const user = await User.findById(req.params.id).select('-password');
-    
-    if (!user) {
-      return res.status(404).json({
-        success: false,
-        message: 'User not found'
-      });
-    }
-
-    res.json({
-      success: true,
-      data: user
-    });
-  } catch (error) {
-    console.error('Get user error:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Server error fetching user'
-    });
-  }
-});
-
-// PUT /api/users/:id/role - Update user role (admin only)
-usersRouter.put('/:id/role', authMiddleware, adminMiddleware, [
-  body('role')
-    .isIn(['user', 'admin'])
-    .withMessage('Invalid role value')
-], async (req, res) => {
-  try {
-    // Check database connection
-    if (mongoose.connection.readyState !== 1) {
-      return res.status(503).json({
-        success: false,
-        message: 'Database connection not available'
-      });
-    }
-
-    // Check for validation errors
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        success: false,
-        message: 'Validation failed',
-        errors: errors.array()
-      });
-    }
-
-    const { id } = req.params;
-    const { role } = req.body;
-
-    // Prevent user from changing their own role
-    if (id === req.user.id) {
-      return res.status(400).json({
-        success: false,
-        message: 'You cannot change your own role'
-      });
-    }
-
-    const user = await User.findByIdAndUpdate(
-      id,
-      { role },
-      { new: true, runValidators: true }
-    ).select('-password');
-
-    if (!user) {
-      return res.status(404).json({
-        success: false,
-        message: 'User not found'
-      });
-    }
-
-    res.json({
-      success: true,
-      message: 'User role updated successfully',
-      data: user
-    });
-
-  } catch (error) {
-    console.error('❌ Error updating user role:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Error updating user role',
-      error: process.env.NODE_ENV === 'development' ? error.message : 'Internal server error'
-    });
-  }
-});
-
-// PUT /api/users/:id/status - Toggle user active status (admin only)
-usersRouter.put('/:id/status', authMiddleware, adminMiddleware, [
-  body('isActive')
-    .isBoolean()
-    .withMessage('isActive must be a boolean value')
-], async (req, res) => {
-  try {
-    // Check database connection
-    if (mongoose.connection.readyState !== 1) {
-      return res.status(503).json({
-        success: false,
-        message: 'Database connection not available'
-      });
-    }
-
-    // Check for validation errors
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        success: false,
-        message: 'Validation failed',
-        errors: errors.array()
-      });
-    }
-
-    const { id } = req.params;
-    const { isActive } = req.body;
-
-    // Prevent user from deactivating themselves
-    if (id === req.user.id && !isActive) {
-      return res.status(400).json({
-        success: false,
-        message: 'You cannot deactivate your own account'
-      });
-    }
-
-    const user = await User.findByIdAndUpdate(
-      id,
-      { isActive },
-      { new: true, runValidators: true }
-    ).select('-password');
-
-    if (!user) {
-      return res.status(404).json({
-        success: false,
-        message: 'User not found'
-      });
-    }
-
-    res.json({
-      success: true,
-      message: `User ${isActive ? 'activated' : 'deactivated'} successfully`,
-      data: user
-    });
-
-  } catch (error) {
-    console.error('❌ Error updating user status:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Error updating user status',
-      error: process.env.NODE_ENV === 'development' ? error.message : 'Internal server error'
-    });
-  }
-});
-
-// Data routes
-const dataRouter = express.Router();
-
-dataRouter.get('/dashboard', (req, res) => {
-  res.json({
-    message: 'Dashboard data',
-    data: {
-      totalUsers: 150,
-      totalOrders: 45,
-      revenue: 12500,
-      growthRate: 15.2
-    }
-  });
-});
-
-dataRouter.get('/analytics', (req, res) => {
-  res.json({
-    message: 'Analytics data',
-    analytics: {
-      pageViews: 1250,
-      uniqueVisitors: 890,
-      bounceRate: 32.5,
-      avgSessionDuration: '2m 45s'
-    }
-  });
-});
-
-// Lead routes (for landing page contact form)
-const leadRouter = express.Router();
-
-// POST /api/leads - Submit a new lead (no authentication required)
-leadRouter.post('/', [
-  body('name')
-    .trim()
-    .isLength({ min: 2, max: 100 })
-    .withMessage('Name must be between 2 and 100 characters'),
-  body('email')
-    .isEmail()
-    .normalizeEmail()
-    .withMessage('Please enter a valid email'),
-  body('company')
-    .optional()
-    .trim()
-    .isLength({ max: 100 })
-    .withMessage('Company name cannot exceed 100 characters'),
-  body('message')
-    .trim()
-    .isLength({ min: 10, max: 1000 })
-    .withMessage('Message must be between 10 and 1000 characters')
-], async (req, res) => {
-  try {
-    // Check database connection
-    if (mongoose.connection.readyState !== 1) {
-      return res.status(503).json({
-        success: false,
-        message: 'Database connection not available'
-      });
-    }
-
-    // Check for validation errors
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        success: false,
-        message: 'Validation failed',
-        errors: errors.array()
-      });
-    }
-
-    const { name, email, company, message } = req.body;
-
-    // Get client IP and user agent for tracking
-    const ipAddress = req.ip || req.connection.remoteAddress || req.socket.remoteAddress || 
-                     (req.connection.socket ? req.connection.socket.remoteAddress : null);
-    const userAgent = req.get('User-Agent');
-
-    // Create new lead
-    const lead = new Lead({
-      name,
-      email,
-      company: company || '',
-      message,
-      ipAddress,
-      userAgent
-    });
-
-    await lead.save();
-
-    console.log('✅ New lead submitted:', { name, email, company });
-
-    res.status(201).json({
-      success: true,
-      message: 'Thank you for your message! We will get back to you soon.',
-      data: {
-        id: lead._id,
-        name: lead.name,
-        email: lead.email,
-        company: lead.company,
-        createdAt: lead.createdAt
-      }
-    });
-
-  } catch (error) {
-    console.error('❌ Error creating lead:', error);
-    res.status(500).json({
-      success: false,
-      message: 'An error occurred while submitting your message. Please try again.',
-      error: process.env.NODE_ENV === 'development' ? error.message : 'Internal server error'
-    });
-  }
-});
-
-// GET /api/leads - Get all leads (admin only)
-leadRouter.get('/', authMiddleware, adminMiddleware, async (req, res) => {
-  try {
-    // Check database connection
-    if (mongoose.connection.readyState !== 1) {
-      return res.status(503).json({
-        success: false,
-        message: 'Database connection not available'
-      });
-    }
-
-    const page = parseInt(req.query.page) || 1;
-    const limit = parseInt(req.query.limit) || 10;
-    const status = req.query.status;
-    const search = req.query.search;
-
-    // Build query
-    let query = {};
-    if (status && status !== 'all') {
-      query.status = status;
-    }
-    if (search) {
-      query.$or = [
-        { name: { $regex: search, $options: 'i' } },
-        { email: { $regex: search, $options: 'i' } },
-        { company: { $regex: search, $options: 'i' } },
-        { message: { $regex: search, $options: 'i' } }
-      ];
-    }
-
-    // Get leads with pagination
-    const leads = await Lead.find(query)
-      .sort({ createdAt: -1 })
-      .skip((page - 1) * limit)
-      .limit(limit);
-
-    // Get total count for pagination
-    const total = await Lead.countDocuments(query);
-
-    // Get status counts for dashboard
-    const statusCounts = await Lead.aggregate([
-      { $group: { _id: '$status', count: { $sum: 1 } } }
-    ]);
-
-    const statusStats = statusCounts.reduce((acc, item) => {
-      acc[item._id] = item.count;
-      return acc;
-    }, {});
-
-    res.json({
-      success: true,
-      data: {
-        leads,
-        pagination: {
-          page,
-          limit,
-          total,
-          pages: Math.ceil(total / limit)
-        },
-        stats: {
-          total,
-          ...statusStats
-        }
-      }
-    });
-
-  } catch (error) {
-    console.error('❌ Error fetching leads:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Error fetching leads',
-      error: process.env.NODE_ENV === 'development' ? error.message : 'Internal server error'
-    });
-  }
-});
-
-// PUT /api/leads/:id/status - Update lead status (admin only)
-leadRouter.put('/:id/status', authMiddleware, adminMiddleware, [
-  body('status')
-    .isIn(['new', 'contacted', 'qualified', 'converted', 'closed'])
-    .withMessage('Invalid status value')
-], async (req, res) => {
-  try {
-    // Check database connection
-    if (mongoose.connection.readyState !== 1) {
-      return res.status(503).json({
-        success: false,
-        message: 'Database connection not available'
-      });
-    }
-
-    // Check for validation errors
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-      return res.status(400).json({
-        success: false,
-        message: 'Validation failed',
-        errors: errors.array()
-      });
-    }
-
-    const { id } = req.params;
-    const { status } = req.body;
-
-    const lead = await Lead.findByIdAndUpdate(
-      id,
-      { status },
-      { new: true, runValidators: true }
-    );
-
-    if (!lead) {
-      return res.status(404).json({
-        success: false,
-        message: 'Lead not found'
-      });
-    }
-
-    res.json({
-      success: true,
-      message: 'Lead status updated successfully',
-      data: lead
-    });
-
-  } catch (error) {
-    console.error('❌ Error updating lead status:', error);
-    res.status(500).json({
-      success: false,
-      message: 'Error updating lead status',
-      error: process.env.NODE_ENV === 'development' ? error.message : 'Internal server error'
-    });
-  }
-});
-
-// Mount sub-routers
-router.use('/users', usersRouter);
-router.use('/auth', authRouter);
-router.use('/data', dataRouter);
-router.use('/leads', leadRouter);
-
-// Test endpoint
-router.get('/test', (req, res) => {
-  res.json({
-    message: 'API test endpoint working!',
-    timestamp: new Date().toISOString(),
-    method: req.method,
-    path: req.path,
-    environment: {
-      nodeEnv: process.env.NODE_ENV,
-      hasJwtSecret: !!process.env.JWT_SECRET,
-      hasMongoUri: !!(process.env.MONGODB_URI || process.env.DA_DATABASE_URL_MONGODB_URI),
-      mongooseReady: mongoose.connection.readyState === 1
-    }
-  });
-});
-
-module.exports = router;
