+++ src/app.js
@@ -1,0 +1,53 @@
+const express = require('express');
+const cors = require('cors');
+const helmet = require('helmet');
+const morgan = require('morgan');
+
+const { getCorsOptions } = require('./config/cors');
+const { errorHandler, notFoundHandler } = require('./middleware/errorHandler');
+const routes = require('./routes');
+
+const app = express();
+
+// Security middleware
+app.use(helmet());
+
+// CORS configuration
+app.use(cors(getCorsOptions()));
+
+// Logging middleware
+app.use(morgan('combined'));
+
+// Body parser middleware
+app.use(express.json({ limit: '10mb' }));
+app.use(express.urlencoded({ extended: true }));
+
+// Handle preflight OPTIONS requests
+app.options('*', (req, res) => {
+  res.header('Access-Control-Allow-Origin', req.headers.origin);
+  res.header('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, OPTIONS');
+  res.header('Access-Control-Allow-Headers', 'Content-Type, Authorization, X-Requested-With, Accept');
+  res.header('Access-Control-Allow-Credentials', 'true');
+  res.sendStatus(200);
+});
+
+// Root endpoint
+app.get('/', (req, res) => {
+  res.json({
+    message: 'DA Admin Backend API',
+    version: '1.0.0',
+    status: 'running',
+    timestamp: new Date().toISOString()
+  });
+});
+
+// API Routes
+app.use('/api', routes);
+
+// Error handling middleware (must be after routes)
+app.use(errorHandler);
+
+// 404 handler (must be last)
+app.use('*', notFoundHandler);
+
+module.exports = app;
