+++ src/middleware/validateRequest.js
@@ -1,0 +1,20 @@
+const { validationResult } = require('express-validator');
+
+/**
+ * Middleware to validate request based on express-validator rules
+ */
+const validateRequest = (req, res, next) => {
+  const errors = validationResult(req);
+  
+  if (!errors.isEmpty()) {
+    return res.status(400).json({
+      success: false,
+      message: 'Validation failed',
+      errors: errors.array()
+    });
+  }
+  
+  next();
+};
+
+module.exports = validateRequest;
