+++ src/utils/responseHandler.js
@@ -1,0 +1,47 @@
+/**
+ * Standard response handlers for API responses
+ */
+
+/**
+ * Send success response
+ */
+const sendSuccess = (res, data, message = 'Success', statusCode = 200) => {
+  return res.status(statusCode).json({
+    success: true,
+    message,
+    ...data
+  });
+};
+
+/**
+ * Send error response
+ */
+const sendError = (res, message, statusCode = 500, errors = null) => {
+  const response = {
+    success: false,
+    message
+  };
+
+  if (errors) {
+    response.errors = errors;
+  }
+
+  return res.status(statusCode).json(response);
+};
+
+/**
+ * Send validation error response
+ */
+const sendValidationError = (res, errors) => {
+  return res.status(400).json({
+    success: false,
+    message: 'Validation failed',
+    errors
+  });
+};
+
+module.exports = {
+  sendSuccess,
+  sendError,
+  sendValidationError
+};
