+++ src/utils/tokenGenerator.js
@@ -1,0 +1,20 @@
+const jwt = require('jsonwebtoken');
+
+/**
+ * Generate JWT token for user authentication
+ * @param {string} userId - User ID to encode in token
+ * @returns {string} JWT token
+ */
+const generateToken = (userId) => {
+  if (!process.env.JWT_SECRET) {
+    throw new Error('JWT_SECRET environment variable is not defined');
+  }
+  
+  return jwt.sign({ userId }, process.env.JWT_SECRET, {
+    expiresIn: '7d' // Token expires in 7 days
+  });
+};
+
+module.exports = {
+  generateToken
+};
