+++ src/validators/authValidators.js
@@ -1,0 +1,36 @@
+const { body } = require('express-validator');
+
+/**
+ * Validation rules for user registration
+ */
+const registerValidation = [
+  body('name')
+    .trim()
+    .isLength({ min: 2, max: 50 })
+    .withMessage('Name must be between 2 and 50 characters'),
+  body('email')
+    .isEmail()
+    .normalizeEmail()
+    .withMessage('Please enter a valid email'),
+  body('password')
+    .isLength({ min: 6 })
+    .withMessage('Password must be at least 6 characters long')
+];
+
+/**
+ * Validation rules for user login
+ */
+const loginValidation = [
+  body('email')
+    .isEmail()
+    .normalizeEmail()
+    .withMessage('Please enter a valid email'),
+  body('password')
+    .notEmpty()
+    .withMessage('Password is required')
+];
+
+module.exports = {
+  registerValidation,
+  loginValidation
+};
